Survive a screen too small for a window, and a resize under the editor
Three ways a resize crashed the reader on a phone, found with an
address sanitizer and random resizes between reading and writing.

A window that does not fit the screen is not opened, but the object
standing for it took whatever window was on top for its own: on a
resize inside the editor the header view's, which the editor's next
resize freed out from under it. The object now knows when it has no
window, and every method that reaches for the record checks first.

Laying the screen out for a new size closes every window, the reader's
two views among them - whoever took the resize, the editor above them
included. The view objects kept their records, painted into them when
the editor returned and closed them again on the next resize. The
layout itself now forgets and re-creates the reader's views, once,
whatever the caller; the editor and the reader no longer each do part
of it.

A window laid out for a taller screen asked for rows past the screen,
and curses read past its buffer copying them: the bottom row is cut to
the screen the way the right column always was, and a window with
nothing left is refused.

And the arithmetic on the unsigned screen size: the editor's row and
column counts wrapped on a screen of fewer than eight rows, and the
info, ask, area list, nodelist, reply list and file list windows placed
or sized themselves with MAXCOL-63, -80 or -36, which wrapped below
that many columns. All signed and clamped now.

+++ golded3/genode.cpp
@@ -278,9 +278,10 @@ void NodelistBrowser::DisplayLine(int line)
 void NodelistBrowser::BuildListString(int line)
 {
 
-    int x1 = (MAXCOL-80)/3;
+    //  MAXCOL is unsigned; below 80 columns the difference wrapped.
+    int x1 = ((int)MAXCOL-80)/3;
     int x2 = x1;
-    int x3 = (MAXCOL-80) - (x1+x2);
+    int x3 = ((int)MAXCOL-80) - (x1+x2);
 
     ftn_nodelist_entry* entryp = entries + (line - 1);
     *entryp = NLP->entry();
